refactor: extract cart context provider into own module which uses reducer
Now methods related to modifying the state of the
cart items is done through the dispatch method of
the reducer inside the new cartProvider component.

Add enum CartActions module for the action names.

Corrected some styling bugs inside the table and
added a prettier empty cart message box.

Changed the tests to use the new provider.