feat: add searching into the 'products' route. add unit tests for user search
Now the search bar for the 'products' page filters the
shown products by matching title. The filtering works
even with partially matching title.

Search bar now shows the results count and a message
for when no products match the search query.

Refactored the 'products' route tests so that the loader
functions that make the search box filtering be tested.

Add dev dependency vitest/ui for seeing the test
results in the browser because checking them in the
console sucked (●'◡'●)