feat: add modules and create cmpnets in regards to getting products by categories
'categories' module contains a method for retrieving
an array of product categories

'products' module has a new method for getting all
the products under a certain category.

Add new loader method to the products route for when
a category is selected.

Create 'categoriesList' component for displaying the
categories as links.