feat: add price range slider and product filters component
Created a price range slider that shows products
that have a price within the selected range.

The price range slider contains a multiRangeSlider
component which holds the two values of the price
range.

The slider values can be changed from two input fields
inside a form in the priceRangeSlider component.