feat: priceRangeSlider correction for input fields and toggle checkbox
Before, the slider defaulted to the respective min/max boundary when user tried to delete the input.

Now, it allows deleting the values min and max and typing the user's own.

The priceRangeSlider checkbox now is using a label which can be clicked to toggle the filter.

Refactored tests for 'products' route to use a more specific test ID for products list.