feat: integrate pagination functionality  into 'products' route
Because no limit or offset is present for the fake API
we simulate it by adding a new 'getPaginatedProducts' method.
It takes a products array and returns only those for
the set page and pageSize.

Include the new page and pagesize filters to the filterNames module.

Add pagination change handlers and components.
Changing pageSize or page calls the setSearchParams method
from the useSearchParams hook.