test: change getDisplayedProducts() method
add GUID for the mock 'product' page to distinguish from 'products'.

Removed the try catch in getDisplayedProducts() in productFiltering.
Now it does not log the error each time and flood the logs.

In 'products.test', when navigating to a selected product now theres a data-testid.

Changed form elements for searching and sorting to not have 'id' attr but 'data-testIds'.